chore(clients/go): move entities to pkg
