+++ clients/go/Makefile
@@ -6,7 +6,7 @@ install-deps:
 .PHONY: test
 test:
 	go test -v ./commands/
-	go test -v ./entities/
+	go test -v ./pkg/entities/
 	go test -v ./worker/
 	go test -v ./tests/
 	go test -v ./zbc/
