+++ clients/go/pkg/entities/entities.go
