+++ clients/go/pkg/entities/job.go
