+++ clients/go/worker/jobDispatcher.go
@@ -15,7 +15,7 @@
 package worker
 
 import (
-	"github.com/zeebe-io/zeebe/clients/go/entities"
+	"github.com/zeebe-io/zeebe/clients/go/pkg/entities"
 	"sync"
 )
