+++ clients/go/worker/jobDispatcher_test.go
@@ -15,14 +15,14 @@
 package worker
 
 import (
-	"github.com/stretchr/testify/suite"
-	"github.com/zeebe-io/zeebe/clients/go/commands"
-	"github.com/zeebe-io/zeebe/clients/go/entities"
-	"github.com/zeebe-io/zeebe/clients/go/internal/pb"
-	"github.com/zeebe-io/zeebe/clients/go/utils"
-	"sync"
-	"testing"
-	"time"
+    "github.com/stretchr/testify/suite"
+    "github.com/zeebe-io/zeebe/clients/go/commands"
+    "github.com/zeebe-io/zeebe/clients/go/internal/pb"
+    "github.com/zeebe-io/zeebe/clients/go/pkg/entities"
+    "github.com/zeebe-io/zeebe/clients/go/utils"
+    "sync"
+    "testing"
+    "time"
 )
 
 type JobDispatcherSuite struct {
