+++ clients/go/worker/jobPoller_test.go
@@ -16,17 +16,17 @@
 package worker
 
 import (
-	"github.com/golang/mock/gomock"
-	"github.com/stretchr/testify/suite"
-	"github.com/zeebe-io/zeebe/clients/go/entities"
-	"github.com/zeebe-io/zeebe/clients/go/internal/mock_pb"
-	"github.com/zeebe-io/zeebe/clients/go/internal/pb"
-	"github.com/zeebe-io/zeebe/clients/go/utils"
-	"io"
-	"math"
-	"sync"
-	"testing"
-	"time"
+    "github.com/golang/mock/gomock"
+    "github.com/stretchr/testify/suite"
+    "github.com/zeebe-io/zeebe/clients/go/internal/mock_pb"
+    "github.com/zeebe-io/zeebe/clients/go/internal/pb"
+    "github.com/zeebe-io/zeebe/clients/go/pkg/entities"
+    "github.com/zeebe-io/zeebe/clients/go/utils"
+    "io"
+    "math"
+    "sync"
+    "testing"
+    "time"
 )
 
 type JobPollerSuite struct {
